requirements.txt
pandas~=2.2.3
streamlit~=1.44.1
python-dotenv~=1.1.0
azure-core~=1.33.0
openai~=1.74.0
fitz~=0.0.1.dev2
requests~=2.32.3
pathlib~=1.0.1
pytesseract~=0.3.13
hdbscan~=0.8.40
umap-learn~=0.5.7
matplotlib~=3.10.3
numpy~=2.2.4
pdf2image~=1.17.0
sentence-transformers~=4.1.0
scikit-learn~=1.6.1
joblib~=1.5.0
torch~=2.7.0